Fixed brittle page serialization. Moved AccountCredentials to web.dto for consistency.
